Move create_sqlalchemy_engine to SqlBackend
