+++ aiida/orm/implementation/sql/backends.py
@@ -11,8 +11,11 @@
 import abc
 from typing import TYPE_CHECKING, Any, Generic, Optional, TypeVar
 
+from sqlalchemy import create_engine
 from sqlalchemy.future.engine import Engine
 from sqlalchemy.orm import Session
+
+from aiida.common import json
 
 from .. import backends
 
@@ -23,6 +26,31 @@ __all__ = ('SqlBackend',)
 
 # The template type for the base ORM model type
 ModelType = TypeVar('ModelType')  # pylint: disable=invalid-name
+
+
+def create_sqlalchemy_engine(profile, **kwargs):
+    """Create SQLAlchemy engine.
+
+    :param kwargs: keyword arguments that will be passed on to `sqlalchemy.create_engine`.
+        See https://docs.sqlalchemy.org/en/13/core/engines.html?highlight=create_engine#sqlalchemy.create_engine for
+        more info.
+    """
+    # The hostname may be `None`, which is a valid value in the case of peer authentication for example. In this case
+    # it should be converted to an empty string, because otherwise the `None` will be converted to string literal "None"
+    hostname = profile.database_hostname or ''
+    separator = ':' if profile.database_port else ''
+
+    engine_url = 'postgresql://{user}:{password}@{hostname}{separator}{port}/{name}'.format(
+        separator=separator,
+        user=profile.database_username,
+        password=profile.database_password,
+        hostname=hostname,
+        port=profile.database_port,
+        name=profile.database_name
+    )
+    return create_engine(
+        engine_url, json_serializer=json.dumps, json_deserializer=json.loads, future=True, encoding='utf-8', **kwargs
+    )
 
 
 class SqlBackend(Generic[ModelType], backends.Backend):
@@ -48,8 +76,6 @@ class SqlBackend(Generic[ModelType], backends.Backend):
 
         :param kwargs: keyword arguments to be passed to the engine
         """
-        from aiida.backends.utils import create_sqlalchemy_engine
-
         if self._engine is None:
             self._engine = create_sqlalchemy_engine(self._profile, **kwargs)
         if self._session is None:
